browser_stack/requirements.txt
Appium-Python-Client==2.5.0